Move nested class cluster out of ClusteringResult
